Feature/simple transactions

+++ internal/v1/categories/services/delete_category.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"financialcontrol/internal/constants"
 	e "financialcontrol/internal/models/errors"
 	"net/http"
 
@@ -12,6 +13,16 @@ func (s CategoriesService) Delete(ctx *gin.Context) (int, []e.ApiError) {
 
 	if len(errs) > 0 {
 		return statusCode, errs
+	}
+
+	hasTransactions, errs := s.repository.HasTransactionsByCategory(ctx, category.ID)
+
+	if len(errs) > 0 {
+		return http.StatusInternalServerError, errs
+	}
+
+	if hasTransactions {
+		return http.StatusBadRequest, []e.ApiError{e.CustomError{Message: constants.CategoryCannotBeDeletedMsg}}
 	}
 
 	errs = s.repository.Delete(ctx, category.ID)
